refactor(test): some refactors in the prject strcture. code and tests are separated
